Standardise the team member commands. Closes #1881
