+++ tests/Bus/Commands/User/AddUserCommandTest.php
@@ -12,23 +12,35 @@
 namespace CachetHQ\Tests\Cachet\Bus\Commands\User;
 
 use AltThree\TestBench\CommandTrait;
-use CachetHQ\Cachet\Bus\Commands\User\InviteTeamMemberCommand;
-use CachetHQ\Cachet\Bus\Handlers\Commands\User\InviteTeamMemberCommandHandler;
+use CachetHQ\Cachet\Bus\Commands\User\AddUserCommand;
+use CachetHQ\Cachet\Bus\Handlers\Commands\User\AddUserCommandHandler;
 use CachetHQ\Tests\Cachet\AbstractTestCase;
 
 /**
- * This is the invite team member command test class.
+ * This is the add user command test class.
  *
+ * @author James Brooks <james@alt-three.com>
  * @author Graham Campbell <graham@alt-three.com>
  */
-class InviteTeamMemberCommandTest extends AbstractTestCase
+class AddUserCommandTest extends AbstractTestCase
 {
     use CommandTrait;
 
     protected function getObjectAndParams()
     {
-        $params = ['emails' => ['foo@example.com']];
-        $object = new InviteTeamMemberCommand($params['emails']);
+        $params = [
+            'username' => 'Test',
+            'password' => 'fooey',
+            'email'    => 'test@test.com',
+            'level'    => 1,
+        ];
+
+        $object = new AddUserCommand(
+            $params['username'],
+            $params['password'],
+            $params['email'],
+            $params['level']
+        );
 
         return compact('params', 'object');
     }
@@ -40,6 +52,6 @@ class InviteTeamMemberCommandTest extends AbstractTestCase
 
     protected function getHandlerClass()
     {
-        return InviteTeamMemberCommandHandler::class;
+        return AddUserCommandHandler::class;
     }
 }
